requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pyreadstat>=1.2.0
python-dotenv>=1.0.0
supabase>=2.0.0
google-generativeai>=0.8.0
